requirements.txt
requests
python-dotenv
pytest
pre-commit
psycopg2-binary
coverage
ruff
pytest-cov
